Move Connect widget tests to own file.
Also only remove the container after every test.

Mount and unmount the widget in the before/afterEach.
