+++ src/__tests__/widget_test.ts
@@ -324,113 +324,3 @@ widgets.forEach((widgetClass) => {
     })
   })
 })
-
-describe("ConnectWidget", () => {
-  describe("handleOAuthRedirect", () => {
-    test("will parse a URL and postmessage error to the widget", async () => {
-      const widget = new ConnectWidget({ url, container })
-
-      await waitFor(() => !!container?.getElementsByTagName("iframe")[0]?.src)
-
-      const iframeElement = container?.getElementsByTagName("iframe")[0]?.contentWindow
-
-      if (!iframeElement) {
-        throw new Error("Unable to find widget iframe")
-      }
-
-      const postMessageSpy = jest.spyOn(iframeElement, "postMessage")
-
-
-      window.dispatchEvent(
-        new MessageEvent("message", {
-          data: {
-              mx: true,
-              type: "mx/client/oauthComplete",
-              metadata: {
-                  url: "com.testapp://oauth_complete?status=error&member_guid=MBR-123&error_reason=CANCELLED"
-              }
-          }
-        })
-      )
-
-      expect(postMessageSpy).toHaveBeenCalledWith(
-        {
-          mx: true,
-          type: "oauthComplete/error",
-          metadata: {
-            member_guid: "MBR-123",
-            error_reason: "CANCELLED"
-          },
-        },
-        widget.targetOrigin,
-      )
-    })
-
-    test("will parse a URL and postmessage success to the widget", async () => {
-      const widget = new ConnectWidget({ url, container })
-
-      await waitFor(() => !!container?.getElementsByTagName("iframe")[0]?.src)
-
-      const iframeElement = container?.getElementsByTagName("iframe")[0]?.contentWindow
-
-      if (!iframeElement) {
-        throw new Error("Unable to find widget iframe")
-      }
-
-      const postMessageSpy = jest.spyOn(iframeElement, "postMessage")
-
-
-      window.dispatchEvent(
-        new MessageEvent("message", {
-          data: {
-              mx: true,
-              type: "mx/client/oauthComplete",
-              metadata: {
-                  url: "com.testapp://oauth_complete?status=success&member_guid=MBR-123"
-              }
-          }
-        })
-      )
-
-      expect(postMessageSpy).toHaveBeenCalledWith(
-        {
-          mx: true,
-          type: "oauthComplete/success",
-          metadata: {
-            member_guid: "MBR-123",
-            error_reason: null,
-          },
-        },
-        widget.targetOrigin,
-      )
-    })
-  })
-
-  describe("post message dispatching", () => {
-    test("message payload is included", () => {
-      const onLoaded = jest.fn()
-      new ConnectWidget({ url, container, onLoaded })
-
-      window.dispatchEvent(
-        new MessageEvent("message", {
-          data: {
-            mx: true,
-            type: "mx/connect/loaded",
-            metadata: {
-              user_guid: "USR-123",
-              session_guid: "SES-123",
-              initial_step: "mfa",
-            },
-          },
-        }),
-      )
-
-      expect(onLoaded).toHaveBeenCalledWith({
-        initial_step: "mfa",
-        session_guid: "SES-123",
-        type: "mx/connect/loaded",
-        user_guid: "USR-123",
-      })
-    })
-  })
-})
